+++ pool_impl/src/linked_pool/GlobalLinkedPool2.h
@@ -113,7 +113,7 @@ void* GlobalLinkedPool::allocate() {
         header->sizeOfObjects = m_sizeOfObjects;
         header->head.next = reinterpret_cast<Node*>(header + 1);
         char* first = reinterpret_cast<char*>(header + 1);
-        for (int i = 0; i < m_poolSize - 1; ++i) {
+        for (size_t i = 0; i < m_poolSize - 1; ++i) {
             Node* node = reinterpret_cast<Node*>(first);
             *node = Node();
             first += m_sizeOfObjects;
