+++ pool_impl/src/linked_pool/LinkedPool.h
@@ -97,7 +97,7 @@ void* LinkedPool<T>::allocate() {
         *header = PoolHeader();
         header->head.next = reinterpret_cast<Node*>(header + 1);
         T* first = reinterpret_cast<T*>(header + 1);
-        for (int i = 0; i < m_poolSize - 1; ++i) {
+        for (size_t i = 0; i < m_poolSize - 1; ++i) {
             Node* node = reinterpret_cast<Node*>(first);
             *node = Node();
             node->next = reinterpret_cast<Node*>(++first);
